+++ include/information/VariableParameters.h
@@ -17,8 +17,7 @@ public:
      *
      * Constructs a %VariableParameters object with default values.
 	 */
-    VariableParameters(const Eigen::VectorXd& popLevels, const Eigen::VectorXd&
-        bridge, const Eigen::VectorXd& hp, const Eigen::VectorXd& l, const
+    VariableParameters(const Eigen::VectorXd& popLevels, const Eigen::VectorXi &bridge, const Eigen::VectorXd& hp, const Eigen::VectorXd& l, const
         Eigen::VectorXd& b, const Eigen::VectorXd& pgr,
         const Eigen::VectorXd& c, const Eigen::VectorXd& csd);
     /**
